make sonarcloud happy

+++ app/src/candid/leb128.c
@@ -47,7 +47,7 @@ parser_error_t decompressLEB128(const uint8_t *input, uint16_t inputSize, uint64
 }
 
 parser_error_t decompressSLEB128(const uint8_t *input, uint16_t inputSize, int64_t *v, uint16_t *bytesConsumed) {
-    unsigned int i = 0;
+    uint16_t i = 0;
 
     *bytesConsumed= 0;
     *v = 0;
@@ -60,7 +60,7 @@ parser_error_t decompressSLEB128(const uint8_t *input, uint16_t inputSize, int64
             if (b == 0 || b == 0x7f) {
                 *bytesConsumed = i + 1;
                 if (b == 0x7f) {
-                    *v |= 1ull << shift;
+                    *v |= 1ULL << shift;
                 }
                 return parser_ok;
             }
@@ -73,7 +73,7 @@ parser_error_t decompressSLEB128(const uint8_t *input, uint16_t inputSize, int64
 
         if ((input[i] & 0x80u) == 0) {
             if (input[i] & 0x40) {
-                *v |= -(1ull << shift);
+                *v |= -(int64_t)(1ULL << shift);
             }
             *bytesConsumed = i + 1;
             return parser_ok;
